adapt imports for wci-browser
