+++ x-pack/solutions/search/packages/workchat/wci-browser/kibana.jsonc
@@ -1,5 +1,5 @@
 {
-  "type": "shared-common",
+  "type": "shared-browser",
   "id": "@kbn/wci-browser",
   "owner": "@elastic/search-kibana",
   "group": "search",
