+++ x-pack/solutions/search/plugins/workchat/workchat_app/public/types.ts
@@ -6,7 +6,7 @@
  */
 
 import type { InferencePublicStart } from '@kbn/inference-plugin/public';
-import type { IntegrationComponentDescriptor } from '@kbn/wci-common';
+import type { IntegrationComponentDescriptor } from '@kbn/wci-browser';
 
 export interface WorkChatAppPluginSetup {
   integrations: {
